the gyroscope values are converted into degrees per second.

+++ src/stm32f4/src/gyroscope.h
@@ -5,10 +5,10 @@
 #include <i2c.h>
 
 typedef struct GyroscopeReading {
-	int16_t x;
-	int16_t y;
-	int16_t z;
-	int16_t gyroscopeTemperature;
+	float x;
+	float y;
+	float z;
+	float gyroscopeTemperature;
 }GyroscopeReading;
 
 struct GyroscopeReading CreateGyroscopeReading();
